% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{group_by_time_interval}
\alias{group_by_time_interval}
\title{Organise results into counts by time interval}
\usage{
group_by_time_interval(pub_dates)
}
\arguments{
\item{pub_dates}{Output from search_term or search_term_by_year}
}
\value{
The search results.
}
\description{
Provide a weekly, monthly or yearly count of the number of publications
}
