% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{plot_time_interval}
\alias{plot_time_interval}
\title{Plot the results of the search}
\usage{
plot_time_interval(x, interval = "Weeks")
}
\arguments{
\item{x}{The results group_by_time_interval}

\item{interval}{The time interval to plot. Default is "Weeks". Choose "Weeks", "Months", "Years".}
}
\value{
A ggplot object.
}
\description{
This function plots the results of the search.
It uses ggplot2 to create a line plot of the number of publications over time.
}
