% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{search_term}
\alias{search_term}
\title{Search for a term in the database.}
\usage{
search_term(term)
}
\arguments{
\item{term}{The term to search for.}
}
\value{
The search results.
}
\description{
This function searches for a given term in the database and returns the results.
}
