% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{search_term_by_year}
\alias{search_term_by_year}
\title{Search for a term in the database, split up by year.}
\usage{
search_term_by_year(term, years)
}
\arguments{
\item{term}{The term to search for.}

\item{years}{The years to search for.}
}
\value{
The search results.
}
\description{
This function searches for a given term in the database and returns the results.
It splits the search into years to avoid hitting the NCBI server limits.
}
